feat(corfu): preview snippet expansion as doc
This shows what expanding the selected snippet would insert in the
buffer. It does so by saving the snippet template as a property of the
candidate, and using it to expand to a separate buffer in the
`:company-doc-buffer` property. Popupinfo calls the value of this
property to get the documentation to display, as a buffer. That's when
we retrieve the template, expand it, fill in any placeholders with
"<number>", insert "<END>" at the final cursor position, and finally
highlight.

Some pieces of misplaced documentation were corrected as well.