feat(corfu): add fish and bash completion support
With this, Fish and Bash completions will show up in shell-mode and in
read-shell-command. This mirrors the Eshell module. Using pcmpl-args is
a possibility, but from what I recall, pcmpl-args is meant to be used on
a per-command basis, and I'm not sure how to handle that in a general
way. This is a lot better than nothing, and it's a start.